refactor(kitchenpos): 도메인별로 재배치
