+++ src/test/java/kitchenpos/menu/acceptance/MenuAcceptanceUtils.java
@@ -1,9 +1,9 @@
-package kitchenpos.acceptance;
+package kitchenpos.menu.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import kitchenpos.dto.MenuRequest;
+import kitchenpos.menu.dto.MenuRequest;
 import org.springframework.http.MediaType;
 
 public class MenuAcceptanceUtils {
