+++ src/test/java/kitchenpos/ordertable/acceptance/TableAcceptanceUtils.java
@@ -1,10 +1,10 @@
-package kitchenpos.acceptance;
+package kitchenpos.ordertable.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.domain.order.OrderTable;
-import kitchenpos.dto.OrderTableRequest;
+import kitchenpos.ordertable.dto.OrderTableRequest;
 import org.springframework.http.MediaType;
 
 public class TableAcceptanceUtils {
@@ -34,20 +34,20 @@ public class TableAcceptanceUtils {
                 .extract();
     }
 
-    public static ExtractableResponse<Response> 주문_테이블_빈_상태_변경_요청(Long orderTableId, OrderTable orderTable) {
+    public static ExtractableResponse<Response> 주문_테이블_빈_상태_변경_요청(Long orderTableId, OrderTableRequest request) {
         return RestAssured
                 .given().log().all()
-                .body(orderTable)
+                .body(request)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when().put("/api/tables/{orderTableId}/empty", orderTableId)
                 .then().log().all()
                 .extract();
     }
 
-    public static ExtractableResponse<Response> 주문_테이블_방문한_손님_수_변경_요청(Long orderTableId, OrderTable orderTable) {
+    public static ExtractableResponse<Response> 주문_테이블_방문한_손님_수_변경_요청(Long orderTableId, OrderTableRequest request) {
         return RestAssured
                 .given().log().all()
-                .body(orderTable)
+                .body(request)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when().put("/api/tables/{orderTableId}/number-of-guests", orderTableId)
                 .then().log().all()
